+++ src/test/java/net/kaczmarzyk/spring/data/jpa/ApplicationWithSARConfiguredWithApplicationContext.java
@@ -38,6 +38,7 @@ import static org.springframework.jmx.support.RegistrationPolicy.IGNORE_EXISTING
 @ComponentScan(basePackages = "net.kaczmarzyk", excludeFilters = {
 		@ComponentScan.Filter(type = ASSIGNABLE_TYPE, value = Application.class),
 		@ComponentScan.Filter(type = ASSIGNABLE_TYPE, value = ApplicationWithConfiguredConversionService.class),
+		@ComponentScan.Filter(type = ASSIGNABLE_TYPE, value = ApplicationWithConfiguredCache.class)
 })
 @EnableJpaRepositories
 @EnableAutoConfiguration