+++ src/test/java/net/kaczmarzyk/spring/data/jpa/domain/NullTest.java
@@ -17,6 +17,8 @@ package net.kaczmarzyk.spring.data.jpa.domain;
 
 import net.kaczmarzyk.spring.data.jpa.Customer;
 import net.kaczmarzyk.spring.data.jpa.IntegrationTestBase;
+import nl.jqno.equalsverifier.EqualsVerifier;
+import nl.jqno.equalsverifier.Warning;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -57,5 +59,13 @@ public class NullTest extends IntegrationTestBase {
 		List<Customer> found = customerRepo.findAll(spec);
 
 		assertThat(found).hasSize(1).containsOnly(bartSimpson);
+	}
+
+	@Test
+	public void equalsAndHashCodeContract() {
+		EqualsVerifier.forClass(Null.class)
+				.usingGetClass()
+				.suppress(Warning.NONFINAL_FIELDS)
+				.verify();
 	}
 }